Cars app. Tweaking for Django version + calendar
Introducing a new calendar for picking the dates to make a reservation.
The goal for this is it simplify and make it more intuitive to the
users. Still needs a bit of work to stop as much wrongdoings by the user
as possible. (Not able to pick overlapping dates, etc.)