Major change hosting to apache nginx. Django Version 1.10.6 (#8)
* Rename project settings folder. Static loc tweak

Changed the position of the static folder and template folder. They will
now sit inside with the project settings.
Updates URLS to fit newer Django recommendations.

* Frontpage App and new packages

Some tweaking to the frontpage app. URLS and forms mainly to fit the
newer DJango version (date input_formats for form validation)

* Cars app. Tweaking for Django version + calendar

Introducing a new calendar for picking the dates to make a reservation.
The goal for this is it simplify and make it more intuitive to the
users. Still needs a bit of work to stop as much wrongdoings by the user
as possible. (Not able to pick overlapping dates, etc.)

* Final Porting Booking/Static/Templates

Final files needed for the porting to work. RUN file for the server,
static files placed appropriately.