404, frontpage bug and fix
Created the 404 file, and it works as inteded when not in debugging
mode. Still needs work on the HTML directly.

Frontpage had a wierd 'bug' where the cursor had pointer in odd places.
This happened because Bootstraps 'checkbox-inline' spanned all the
checkboxes, making it think it was one huge field with checkboxes.